+++ src/MaximaCommandMenus.cpp
@@ -1,0 +1,86 @@
+// -*- mode: c++; c-file-style: "linux"; c-basic-offset: 2; indent-tabs-mode: nil -*-
+//
+//  Copyright (C) 2026 Gunter Königsmann <wxMaxima@physikbuch.de>
+//
+//  This program is free software; you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation; either version 2 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program; if not, write to the Free Software
+//  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
+//
+//  SPDX-License-Identifier: GPL-2.0+
+
+/*! \file
+  Implementation of the menu-command handlers extracted from wxMaxima - see
+  MaximaCommandMenus.h.
+*/
+
+#include "MaximaCommandMenus.h"
+
+#include "wxMaxima.h"
+#include "EventIDs.h"
+#include <wx/windowptr.h>
+#include "wizards/Plot2dWiz.h"
+#include "wizards/Plot3dWiz.h"
+#include "wizards/PlotFormatWiz.h"
+
+void MaximaCommandMenus::PlotMenu(wxCommandEvent &event) {
+  if (!m_wxMaxima.GetWorksheet())
+    return;
+  m_wxMaxima.GetWorksheet()->CloseAutoCompletePopup();
+
+  wxString expr = m_wxMaxima.GetDefaultEntry();
+  if ((event.GetId() == EventIDs::button_plot3) ||
+      (event.GetId() == EventIDs::gp_plot3)) {
+    wxWindowPtr<Plot3DWiz> wiz(
+      new Plot3DWiz(&m_wxMaxima, -1, &m_wxMaxima.m_configuration, _("Plot 3D")));
+    wiz->SetValue(expr);
+    // wiz->Centre(wxBOTH);
+    wiz->ShowWindowModalThenDo([this, wiz](int retcode) {
+      if (retcode == wxID_OK) {
+        wxString val = wiz->GetValue();
+        m_wxMaxima.MenuCommand(val);
+      }
+    });
+  } else if (event.GetId() == EventIDs::menu_animationautostart) {
+    if (event.IsChecked())
+      m_wxMaxima.MenuCommand(wxS("wxanimate_autoplay:true$"));
+    else
+      m_wxMaxima.MenuCommand(wxS("wxanimate_autoplay:false$"));
+  } else if (event.GetId() == EventIDs::menu_animationframerate) {
+    m_wxMaxima.CommandWiz(_("Enter new animation frame rate [Hz, integer]:"),
+                          wxEmptyString, wxEmptyString,
+                          wxS("wxanimate_framerate : #1#$"), _("Frame rate"),
+                          wxS("%"), wxEmptyString);
+  } else if ((event.GetId() == EventIDs::button_plot2) ||
+             (event.GetId() == EventIDs::gp_plot2)) {
+    wxWindowPtr<Plot2DWiz> wiz(
+      new Plot2DWiz(&m_wxMaxima, -1, &m_wxMaxima.m_configuration, _("Plot 2D")));
+    wiz->SetValue(expr);
+    // wiz->Centre(wxBOTH);
+    wiz->ShowWindowModalThenDo([this, wiz](int retcode) {
+      if (retcode == wxID_OK) {
+        wxString val = wiz->GetValue();
+        m_wxMaxima.MenuCommand(val);
+      }
+    });
+  } else if (event.GetId() == EventIDs::menu_plot_format) {
+    wxWindowPtr<PlotFormatWiz> wiz(new PlotFormatWiz(
+      &m_wxMaxima, -1, &m_wxMaxima.m_configuration, _("Plot format")));
+    wiz->Center(wxBOTH);
+    wiz->ShowWindowModalThenDo([this, wiz](int retcode) {
+      if (retcode == wxID_OK) {
+        m_wxMaxima.MenuCommand(wiz->GetValue());
+      }
+    });
+  }
+}